refactor(billing): extract transaction reporting into its own service
Move getCustomerTransactions and the CSV export stream out of StripeService
into a dedicated TransactionReportingService, and delegate to it from the
controller. Behavior-preserving: the reporting tests move alongside the new
service and StripeService sheds its CSV/reporting dependencies.
